% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{annotateExpectedDonors}
\alias{annotateExpectedDonors}
\title{Annotation of whether or not donors were expected to show up as present in
rollcall}
\usage{
annotateExpectedDonors(rollcallDf, expectedDonors)
}
\arguments{
\item{rollcallDf}{dataframe containing rollcall results}

\item{expectedDonors}{character vector containing subset of donors from 
the VCF with which rollcall was run.}
}
\value{
list of rollcall dataframes with added `Expected` column
to specify whether or not the sample was in the list of expected donors
}
\description{
Annotation of whether or not donors were expected to show up as present in
rollcall
}
