% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{annotateOutliers}
\alias{annotateOutliers}
\title{Annotate any "intruders" and "no-shows" in dataset}
\usage{
annotateOutliers(rollcallDf)
}
\arguments{
\item{rollcallDf}{rollcallDf with "Expected" column stating whether
or not a sample was expected to be found in the sample}
}
\value{
rollcallDf with new column `isOutlier` to specify outliers
}
\description{
Annotate any "intruders" and "no-shows" in dataset
}
