% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{annotatePossibleDonors}
\alias{annotatePossibleDonors}
\title{Annotation of possible donors within a given superset for use with census
validation experiments}
\usage{
annotatePossibleDonors(rollcallDf, possibleDonors)
}
\arguments{
\item{rollcallDf}{dataframe containing rollcall results with an Expected
column as added by the function `annotateExpectedDonors`}

\item{possibleDonors}{list of somewhat likely donors (for identifying easy
typos or swaps that occurred within a batch of samples. Used specifically for
census validation experiments.}
}
\description{
Annotation of possible donors within a given superset for use with census
validation experiments
}
