% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateDoubletDetection.R
\name{evaluateBestPairPvalue}
\alias{evaluateBestPairPvalue}
\title{Evaluate Doublet rate in the context of the 2nd best donor p-value.}
\usage{
evaluateBestPairPvalue(
  doubletLikelihoodFile,
  likelihoodSummaryFile,
  expectedSamplesFile,
  pvalThresholdDoublet = 0.9,
  numQuantiles = 10
)
}
\arguments{
\item{doubletLikelihoodFile}{The dropulation doublet likelihood file.}

\item{likelihoodSummaryFile}{The dropulation single donor likelihood file.}

\item{expectedSamplesFile}{A file containing the expected list of donors in the experiment.  (optional)}

\item{pvalThresholdDoublet}{The minimum doublet pvalue to call a doublet}

\item{numQuantiles}{How many bins to split data up into for the median to besst likelihood plot.}
}
\description{
Additional metric in doublet detection evaluates the probability that the best donor pair is confident compared to
all other possible donor pairs.
}
