% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateDoubletDetection.R
\name{evaluateDoubletRate}
\alias{evaluateDoubletRate}
\title{Evaluate Doublet Rate}
\usage{
evaluateDoubletRate(
  doubletLikelihoodFile,
  likelihoodSummaryFile,
  expectedSamplesFile,
  pvalThresholdDoublet = 0.9,
  numQuantiles = 10
)
}
\arguments{
\item{doubletLikelihoodFile}{The dropulation doublet likelihood file.}

\item{likelihoodSummaryFile}{The dropulation single donor likelihood file.}

\item{expectedSamplesFile}{A file containing the expected list of donors in the experiment.  (optional)}

\item{pvalThresholdDoublet}{The minimum doublet pvalue to call a doublet}

\item{numQuantiles}{How many bins to split data up into for the median to besst likelihood plot.}
}
\value{
A dataframe containing summary info about the fraction of doublets and enrichment.
}
\description{
The calculation of the doublet rate can be conflated by cell selection, ambient RNA, and the types
of expression used (intronic/coding/etc).  This analysis may help inform what the true doublet rate
is, and if the calculated doublet rate is under/over inflated.
}
