% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{findREPIRVThreshold}
\alias{findREPIRVThreshold}
\title{Determine what REP_IRV threshold is useful to find expected/unexpected donors.}
\usage{
findREPIRVThreshold(
  rollCallFile,
  donorListFile,
  minNumSNPS = 1000,
  alpha = 0.95,
  expName = ""
)
}
\arguments{
\item{rollCallFile}{The Roll Call output file}

\item{donorListFile}{A list of donors expected in the data set.}

\item{minNumSNPS}{The minimum number of SNPs to consider a roll call result to be well powered.}

\item{alpha}{At what pvalue do we believe the classification of donors in the pool is correct?}

\item{expName}{The name of the experiment.}
}
\value{
The REP_IRV threshold above which donors are present in the pool.  Return NA if the threshold can not be determined.
}
\description{
Trains a k=2 gaussian mixture model on the expected donors and unexpcted donors REP_IRV values, 
then tries to find the best separator between the two gaussians to use as a threshold to find 
unexpected donors.
}
