% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{loadRollcallFilesWithIds}
\alias{loadRollcallFilesWithIds}
\title{Loading of rollcall file and addition of column specifying an ID for it}
\usage{
loadRollcallFilesWithIds(rollcallFiles, ids)
}
\arguments{
\item{rollcallFiles}{character vector of rollcall files}

\item{ids}{character vector of ids to assign to data from each rollcall file,
added as a new column ID to the parsed contents of each rollcall file}
}
\description{
Loading of rollcall file and addition of column specifying an ID for it
}
