% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DonorAssignmentStandardAnalysis.R
\name{plotConditionalDoubletRate}
\alias{plotConditionalDoubletRate}
\title{Compute and plot the conditional doublet rate.}
\usage{
plotConditionalDoubletRate(cellDF)
}
\arguments{
\item{cellDF}{dataframe of cell statistics and labels generated by getSingletDoubletDF}
}
\description{
Does the doublet rate change with the number of UMIs?
It might be over-estimated for cells with few UMIs.
From internal synthetic mixture experiments, the doublet false positive rate tends 
to be higher for cells with fewer than ~ 100 or ~ 150 informative UMIs.
}
