% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DonorAssignmentStandardAnalysis.R
\name{plotFractionConfidentDoubletsFromSingleLikelihoodFit}
\alias{plotFractionConfidentDoubletsFromSingleLikelihoodFit}
\title{Plot how well separated the singlets and doublets are.}
\usage{
plotFractionConfidentDoubletsFromSingleLikelihoodFit(cellDF)
}
\arguments{
\item{cellDF}{dataframe of cell statistics and labels generated by getSingletDoubletDF}
}
\description{
Fits the singlet likelihoods to a constrained B-spline model and then plots the residuals of the doublets.
The residuals are the difference between the observed likelihood and the expected likelihood based on the singlet model.
}
