% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateUnguidedLikelihoodAnalysis.R
\name{plotFractionImpossibleAllelesFromDoublets}
\alias{plotFractionImpossibleAllelesFromDoublets}
\title{Plots that describe the errors in the doublet donor assignment}
\usage{
plotFractionImpossibleAllelesFromDoublets(doubletFile, expName)
}
\arguments{
\item{expName}{The experiment name.}
}
\description{
Given the doublet analysis, partition cell barcodes into confident singlets,
confident doublets, and ambient doublets.
For each cell, calculate the fraction of "impossible alleles" - that is, 
alleles that are observed that could not have been generated by the donor.
For single cells, this is an approximation of the error rate, though perhaps 
not the most sensitive one.
}
