% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateUnguidedLikelihoodAnalysis.R
\name{plotPopulationLikelihood}
\alias{plotPopulationLikelihood}
\title{Plot the best likelihood per donor compared to the population likelihood score}
\usage{
plotPopulationLikelihood(
  likelihoodSummaryFile,
  doubletLikelihoodFile,
  expectedSamplesFile,
  alpha = 0.05,
  alphaDoublet = 0.9,
  outPDF = NULL
)
}
\arguments{
\item{likelihoodSummaryFile}{The dropulation single donor likelihood file.}

\item{doubletLikelihoodFile}{The dropulation doublet likelihood file.}

\item{expectedSamplesFile}{A file containing the expected list of donors in the experiment.  (optional)}

\item{alpha}{The p-value threshold for doublets.}

\item{outPDF}{A file to write a PDF with this plot to.  If set to null, no PDF file is emitted.}
}
\value{
The single donor likelihood data with categorization in the status column.
}
\description{
Plot the best likelihood per donor compared to the population likelihood score
}
