% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{plotRollcalls}
\alias{plotRollcalls}
\title{Plotting of rollcall data}
\usage{
plotRollcalls(
  preppedRollcallData,
  labelOutliers = F,
  logScale = F,
  jitterHeight = 0.2,
  pointsAlpha = 0.5,
  labelAlpha = 0.7,
  labelSize = 5,
  flipAxes = FALSE,
  colors = c("black", "#d18017", "red")
)
}
\arguments{
\item{preppedRollcallData}{dataframe containg rollcall data generated using
\code{\link{prepareRollcallData}} or \code{\link{prepareCenValRollcallData}}}

\item{labelOutliers}{logical flag set to label any "intruders" or "no-shows"}

\item{logScale}{logical flag set to log-transform data}

\item{flipAxes}{logical flag to orient x/y axes. Defaults to FALSE, with REP_IRVs on the x-axis}

\item{colors}{colors for different classes (Expected, [Possible], Unexpected)}
}
\value{
ggplot object of plot for given rollcall data
}
\description{
Plotting of rollcall data
}
