% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{plotUnexpectedDonors}
\alias{plotUnexpectedDonors}
\title{Plot the donors that appear in roll call above some threshold value but are not in the donor list.}
\usage{
plotUnexpectedDonors(
  rollCallFile,
  donorListFile,
  repIrvThreshold = NULL,
  expName,
  minNumSNPS = 1000
)
}
\arguments{
\item{rollCallFile}{The roll call file}

\item{donorListFile}{A list of donors expected in the experiment}

\item{repIrvThreshold}{The REP_IRV threshold at which to decide a donor is in the experiment}

\item{expName}{The name of the experiment for plots.}

\item{minNumSNPS}{The minimum number of SNPs to consider a donor in roll call.}
}
\description{
Plot the donors that appear in roll call above some threshold value but are not in the donor list.
}
