% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{predictClassByMedianMeasure}
\alias{predictClassByMedianMeasure}
\title{Helper function to \code{\link{annotateOutliers}}}
\usage{
predictClassByMedianMeasure(data, classCol, measureCol)
}
\arguments{
\item{classCol}{character vector specifying column of `data` used to label class of each point}

\item{measureCol}{character vector specifying column of `data` used as measurement}
}
\value{
vector of type `class(data[[classCol]])` with predicted classes
according to median values of `data[[measureCol]]` per class
}
\description{
Used for predicting misclassified points
}
