% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{prepareCenValRollcallData}
\alias{prepareCenValRollcallData}
\title{Preparation of rollcall data for a CensusValidation experiment}
\usage{
prepareCenValRollcallData(rollcallFiles, donorFiles, supersetFile, ids = NULL)
}
\arguments{
\item{rollcallFiles}{character vector of rollcall file paths}

\item{donorFiles}{character vector of line-separated donor files
containing a subset of donors from the VCF with which rollcall
was run.}

\item{supersetFile}{character file path of newline separated file containing
set of donors run in a larger batch for census validation.}

\item{ids}{character vector of ids corresponding to each experiment}
}
\value{
list of pepared rollcall dataframes
}
\description{
Preparation of rollcall data for a CensusValidation experiment
}
