% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{prepareRollcallData}
\alias{prepareRollcallData}
\title{Preparation of rollcall data to be done before plotting}
\usage{
prepareRollcallData(rollcallFiles, donorFiles, ids = NULL)
}
\arguments{
\item{rollcallFiles}{character vector of rollcall file paths}

\item{donorFiles}{character vector of line-separated donor files
containing a subset of donors from the VCF with which rollcall
was run.}

\item{ids}{labels given to eah rollcall dataset. Added as ID column.}
}
\value{
list of pepared rollcall dataframes
}
\description{
Preparation of rollcall data to be done before plotting
}
