% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateUnguidedLikelihoodAnalysis.R
\name{runLikelihoodStandardAnalysis}
\alias{runLikelihoodStandardAnalysis}
\title{Standard set of analysis for Dropulation runs.}
\usage{
runLikelihoodStandardAnalysis(
  expName,
  likelihoodSummaryFile,
  doubletLikelihoodFile,
  dgeSummaryFile,
  readsPerCellFile = NULL,
  censusFile = NULL,
  expectedSamplesFile = NULL,
  outFileLikelyDonors = NULL,
  outDonorToCellMap = NULL,
  outPDF = NULL,
  outSummaryStatsFile = NULL,
  minimumFractionDonor = 0.002,
  alpha = 0.05,
  minNumUMIs = 0,
  dgeFile = NULL,
  outMetaCellFile = NULL,
  selectedCellsForMetaCellFile = NULL,
  outCellBarcodesFile = NULL
)
}
\arguments{
\item{expName}{The experiment name.}

\item{likelihoodSummaryFile}{The dropulation single donor likelihood file.}

\item{doubletLikelihoodFile}{The dropulation doublet likelihood file.}

\item{dgeSummaryFile}{The DGE summary file.}

\item{readsPerCellFile}{The reads per cell file. (optional)}

\item{censusFile}{The census results file. (optional)}

\item{expectedSamplesFile}{A file containing the expected list of donors in the experiment.  (optional)}

\item{outFileLikelyDonors}{Output of which donors are in the experiment based on the Dropulation results.}

\item{outDonorToCellMap}{A map from each cell barcode to the best donor}

\item{outPDF}{The output PDF file. (optional)}

\item{outSummaryStatsFile}{The output summary statistics file. (optional)}

\item{minimumFractionDonor}{To determine which donors are in the data, what fraction of cells must contain a donor for the donor to be included in the output.  This controls for a low level of mis-calls of donor identity.}

\item{alpha}{The p-value threshold for doublets.}

\item{minNumUMIs}{Filter cells with fewer than this number of UMIs. (optional)}

\item{dgeFile}{The input DGE file to generate meta cell output. (optional)}

\item{outMetaCellFile}{Sums the expression across cells of each donor to generate a new expression file for donors instead of cells. (optional)}

\item{selectedCellsForMetaCellFile}{Limits meta cell expression to a subset of donors.}

\item{outCellBarcodesFile}{cell barcodes for donors in the experiment (optional)}
}
\description{
Standard set of analysis for Dropulation runs.
}
