% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{validateAndExtractIds}
\alias{validateAndExtractIds}
\title{Validation of ids variable if provided, extraction of ids if NULL}
\usage{
validateAndExtractIds(rollcallFiles, ids)
}
\description{
Validation of ids variable if provided, extraction of ids if NULL
}
