% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{validateFilepaths}
\alias{validateFilepaths}
\title{Validation of the existence of files}
\usage{
validateFilepaths(filepaths)
}
\arguments{
\item{filepaths}{character vector of filepaths to validate}
}
\description{
Validation of the existence of files
}
