% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{validateRollcallFiles}
\alias{validateRollcallFiles}
\title{Validation of files needed to prepare a rollcall dataset}
\usage{
validateRollcallFiles(rollcallFiles, donorFiles)
}
\arguments{
\item{rollcallFiles}{character vector of rollcall files to validate}

\item{donorFiles}{character vector of donor files to validate}
}
\description{
Validation of files needed to prepare a rollcall dataset
}
