% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DropulationCensus_StandardAnalysis.R
\name{validateSupersetFile}
\alias{validateSupersetFile}
\title{Validation of supersetFile for census validation experiments}
\usage{
validateSupersetFile(supersetFile)
}
\arguments{
\item{supersetFile}{file containing superset used in census validation experiments}
}
\description{
Validation of supersetFile for census validation experiments
}
